+++ packages/models/lib/main.reflectable.dart
@@ -909,9 +909,10 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix26.IconIntrinsicStates() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix26.IconIntrinsicStates() : null,
               r'create': (bool b) =>
-                  () => b ? prefix26.IconIntrinsicStates.create() : null
+                  () => b ? const prefix26.IconIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -960,9 +961,9 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix26.ListViewIntrinsicStates() : null,
-              r'create': (bool b) =>
-                  () => b ? prefix26.ListViewIntrinsicStates.create() : null
+                  () => b ? const prefix26.ListViewIntrinsicStates() : null,
+              r'create': (bool b) => () =>
+                  b ? const prefix26.ListViewIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1038,9 +1039,9 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {
               r'': (bool b) =>
-                  () => b ? prefix26.ScaffoldIntrinsicStates() : null,
-              r'create': (bool b) =>
-                  () => b ? prefix26.ScaffoldIntrinsicStates.create() : null
+                  () => b ? const prefix26.ScaffoldIntrinsicStates() : null,
+              r'create': (bool b) => () =>
+                  b ? const prefix26.ScaffoldIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1063,9 +1064,10 @@ final _data = <r.Reflectable, r.ReflectorData>{
             {},
             {},
             {
-              r'': (bool b) => () => b ? prefix26.StackIntrinsicStates() : null,
+              r'': (bool b) =>
+                  () => b ? const prefix26.StackIntrinsicStates() : null,
               r'create': (bool b) =>
-                  () => b ? prefix26.StackIntrinsicStates.create() : null
+                  () => b ? const prefix26.StackIntrinsicStates.create() : null
             },
             -1,
             -1,
@@ -1210,7 +1212,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
       null, null, <Type>[prefix25.DynamicIntrinsicState], 1, {}, {}, null, [])
 };
 
-final _memberSymbolMap = null;
+const _memberSymbolMap = null;
 
 void initializeReflectable() {
   r.data = _data;